+++ srsenb/src/metrics_stdout.cc
@@ -150,7 +150,7 @@ std::string metrics_stdout::float_to_string(float f, int digits)
     precision = digits-1;
   }
   else {
-    precision = digits - (int)(log10(fabs(f))-2*DBL_EPSILON);
+    precision = digits - (int)(log10f(fabs(f)) - 2 * DBL_EPSILON);
   }
   os << std::setw(6) << std::fixed << std::setprecision(precision) << f;
   return os.str();
@@ -158,7 +158,7 @@ std::string metrics_stdout::float_to_string(float f, int digits)
 
 std::string metrics_stdout::float_to_eng_string(float f, int digits)
 {
-  const int degree = (f == 0.0) ? 0 : lrint( floor( log10( fabs( f ) ) / 3) );
+  const int degree = (f == 0.0) ? 0 : lrint(floor(log10f(fabs(f)) / 3));
 
   std::string factor;
